Extracted PerFileScoringTask from AnalysisPipeline.Run
* New OspreySharp/Tasks/PerFileScoringTask.cs owns Stage 1
  (LoadLibrary + GenerateDecoys), Stage 2-4 per-file ProcessFile
  loop, the --join-only parquet load branch, and the
  --join-at-pass=2 1st-pass FDR sidecar overlay
* Outputs (FullLibrary, LibraryById, PerFileEntries,
  PerFileCalibrations, PerFileParquetPaths) flow to FirstJoinTask
  through instance properties on the task
* Early-exit paths -- empty library (error), missing/invalid
  1st-pass sidecar on --join-at-pass=2 (error), no scored entries
  (warning), --no-join completion (success) -- return false from
  Run with ctx.ExitCode set; AnalysisPipeline.Run propagates the
  code straight through
* ProcessFile becomes internal so the new task can drive it; method
  body stays on AnalysisPipeline for now
* AnalysisPipeline.Run drops 416 lines of inline orchestration in
  favor of 25 lines of task construction + dispatch. With this
  the Phase A 4-task structure is complete; AnalysisPipeline.Run
  is a thin driver that walks the workflow.html HPC fan-out / join
  boundaries
* Stellar 3-file snapshot regression PASS at every stage

See ai/todos/active/TODO-20260509_osprey_pipeline_tasks.md
