+++ pwiz_tools/OspreySharp/OspreySharp/AnalysisPipeline.cs
@@ -109,422 +109,31 @@ namespace pwiz.OspreySharp
 
             try
             {
-                // Stage 1: Load library + generate decoys
-                var swLibrary = Stopwatch.StartNew();
-                var library = LoadLibrary(config);
-                if (library == null || library.Count == 0)
-                {
-                    LogError("Library is empty after loading");
-                    return 1;
-                }
+                // Per-file scoring phase (Stages 1-4): library load
+                // + decoys + per-file ProcessFile loop. Driven by
+                // PerFileScoringTask; outputs (FullLibrary,
+                // LibraryById, PerFileEntries, PerFileCalibrations,
+                // PerFileParquetPaths) flow back through instance
+                // properties on the task. Returning false here
+                // means the library was empty (error), the
+                // --join-at-pass=2 sidecar load failed (error), or
+                // the --no-join CLI flag wants us to stop after the
+                // per-file fan-out (success). ctx.ExitCode is the
+                // process exit code AnalysisPipeline.Run should
+                // propagate in any of those cases.
+                var perFileScoringTask = new PerFileScoring.PerFileScoringTask(this);
+                var perFileScoringCtx = new Tasks.PipelineContext(
+                    config, LogInfo, LogWarning, LogError);
+                var perFileScoringPipeline = new Tasks.Pipeline(
+                    new Tasks.OspreyTask[] { perFileScoringTask });
+                if (!perFileScoringPipeline.Execute(perFileScoringCtx))
+                    return perFileScoringCtx.ExitCode;
 
-                int nLibraryTargets = 0;
-                foreach (var entry in library)
-                {
-                    if (!entry.IsDecoy)
-                        nLibraryTargets++;
-                }
-                double libLoadSec = swLibrary.Elapsed.TotalSeconds;
-                LogInfo(string.Format("[COUNT] Library targets loaded: {0}", nLibraryTargets));
-
-                List<LibraryEntry> decoys;
-                if (config.ExpectReconciledInput)
-                {
-                    // --join-at-pass=2: decoy LibraryEntries are unused
-                    // downstream. The reconciled parquet already carries
-                    // both target and decoy FDR rows with their stage-1-4
-                    // scores; Stage 5 is skipped, Stage 6 is skipped, and
-                    // the protein-parsimony / blib write paths both filter
-                    // on `entry.IsDecoy` (only target LibraryEntries get
-                    // looked up by entry_id). Skipping the rebuild saves
-                    // ~45s on Astral 1-file (BuildDecoyFromSequence +
-                    // RecalculateFragments dominated the Stage 7+blib
-                    // hotspot list). dotTrace OWN-time on Astral 1-file
-                    // Stage 7 cs run before this fix:
-                    //   BuildDecoyFromSequence  total=45665 ms (89% wall)
-                    //   GenerateDecoys.<>b__0   total=46792 ms
-                    decoys = new List<LibraryEntry>();
-                }
-                else if (!config.DecoysInLibrary)
-                {
-                    List<LibraryEntry> validTargets;
-                    decoys = GenerateDecoys(library, config, out validTargets);
-                    library = validTargets;
-                }
-                else
-                {
-                    decoys = new List<LibraryEntry>();
-                }
-                swLibrary.Stop();
-                double totalSec = swLibrary.Elapsed.TotalSeconds;
-                LogInfo(string.Format("[TIMING] Library loading + decoys: {0:F1}s (load: {1:F1}s, decoys: {2:F1}s)",
-                    totalSec, libLoadSec, totalSec - libLoadSec));
-
-                LogInfo(string.Format("[COUNT] Library decoys generated: {0}", decoys.Count));
-
-                var fullLibrary = new List<LibraryEntry>(library.Count + decoys.Count);
-                fullLibrary.AddRange(library);
-                fullLibrary.AddRange(decoys);
-
-                LogInfo(string.Format("Full library: {0} entries ({1} targets + {2} decoys)",
-                    fullLibrary.Count, library.Count, decoys.Count));
-                LogInfo(string.Format("[COUNT] Full library: {0} ({1} targets + {2} decoys)",
-                    fullLibrary.Count, library.Count, decoys.Count));
-
-                // Count entries with few fragments (diagnostic for entry count parity)
-                int nZeroFrag = 0, nOneFrag = 0, nTwoFrag = 0;
-                foreach (var entry in fullLibrary)
-                {
-                    int fc = entry.Fragments != null ? entry.Fragments.Count : 0;
-                    if (fc == 0)
-                        nZeroFrag++;
-                    else if (fc == 1)
-                        nOneFrag++;
-                    else if (fc == 2)
-                        nTwoFrag++;
-                }
-                if (nZeroFrag + nOneFrag + nTwoFrag > 0)
-                    LogInfo(string.Format("[COUNT] Entries with <3 fragments: {0} (0={1}, 1={2}, 2={3})",
-                        nZeroFrag + nOneFrag + nTwoFrag, nZeroFrag, nOneFrag, nTwoFrag));
-
-                // Build library lookup by ID for fast access
-                var libraryById = new Dictionary<uint, LibraryEntry>(fullLibrary.Count);
-                foreach (var entry in fullLibrary)
-                    libraryById[entry.Id] = entry;
-
-                // Stage 2-4: Per-file calibration + coelution scoring
-                // Process files in parallel when multiple files are provided.
-                var perFileEntries = new List<KeyValuePair<string, List<FdrEntry>>>();
-                // Per-file RT calibration handles harvested by ProcessFile so
-                // Stage 6 reconciliation has the live RTCalibration objects.
-                var perFileCalibrations = new ConcurrentDictionary<string, RTCalibration>();
-                // fileName -> .scores.parquet path. Populated by --join-only
-                // (which already knows the paths) so Stage 6 reconciliation
-                // can lazily load CWT candidates per file via
-                // ParquetScoreCache.LoadCwtCandidatesFromParquet.
-                var perFileParquetPaths = new Dictionary<string, string>();
-
-                bool joinOnly = config.InputScores != null && config.InputScores.Count > 0;
-                int nFiles = joinOnly ? config.InputScores.Count : config.InputFiles.Count;
-
-                // In --input-scores mode without explicit -i, synthesize
-                // InputFiles from the parquet stems so downstream code
-                // (Stage 6 rescore's fileNameToIdx in particular) can map
-                // each file_name back to a real (synthetic) input path.
-                // Mirrors RescoreWorker.RunWorker's synthesis at
-                // AnalysisPipeline.Stage6Rescore.cs:104 and Rust's
-                // run_analysis idempotent synthesis at
-                // pipeline.rs ~line 3144.
-                if (joinOnly && (config.InputFiles == null || config.InputFiles.Count == 0))
-                {
-                    var synthetic = new List<string>(config.InputScores.Count);
-                    foreach (var p in config.InputScores)
-                        synthetic.Add(RescoreHydration.SyntheticInputFromParquet(p));
-                    config.InputFiles = synthetic;
-                }
-
-                // Pre-compute the parquet footer metadata ONCE, against the
-                // unmutated outer config. ProcessFile clones the config per
-                // file and mutates FragmentTolerance during MS2 calibration,
-                // so reading config.SearchParameterHash() inside ProcessFile
-                // would produce a hash that the join-only validator would
-                // not recognize. Built unconditionally (in any non-joinOnly
-                // mode) because Stage 6 reconciliation needs the per-file
-                // .scores.parquet on disk to lazily load CWT candidates —
-                // matches Rust's end-to-end behavior, which always writes
-                // the parquet sidecar regardless of --no-join.
-                Dictionary<string, string> parquetFooterMetadata = null;
-                if (!joinOnly)
-                {
-                    parquetFooterMetadata = new Dictionary<string, string>
-                    {
-                        { "osprey.version", Program.VERSION },
-                        { "osprey.search_hash", config.SearchParameterHash() },
-                        { "osprey.library_hash", config.LibraryIdentityHash() },
-                        { "osprey.reconciled", "false" },
-                    };
-                }
-
-                // File-level parallelism is configurable via
-                // OSPREY_MAX_PARALLEL_FILES:
-                //   1      = strictly sequential (one file at a time)
-                //   N > 1  = up to N files concurrently via Parallel.For
-                //            with MaxDegreeOfParallelism = N
-                //   unset/<=0 = Parallel.For default (all files at once,
-                //            bounded by the TPL thread pool)
-                // Memory-critical 3-file Astral runs on a 64 GB machine
-                // can cap this at 1 or 2 to keep the per-file pool +
-                // spectra under budget; Stellar stays on default for
-                // max throughput.
-                int maxParallelFiles = OspreyEnvironment.MaxParallelFiles;
-
-                // Determine how many files will actually run concurrently so
-                // ProcessFile can divide the inner main-search thread budget
-                // and avoid oversubscription (see notes there). Stored on
-                // the config so the per-file clones inherit it.
-                if (nFiles == 1 || maxParallelFiles == 1)
-                    config.EffectiveFileParallelism = 1;
-                else if (maxParallelFiles > 1)
-                    config.EffectiveFileParallelism = Math.Min(maxParallelFiles, nFiles);
-                else
-                    config.EffectiveFileParallelism = Math.Min(nFiles, Environment.ProcessorCount);
-
-                var swAllFiles = Stopwatch.StartNew();
-                if (joinOnly)
-                {
-                    // --join-only: load per-file FdrEntry stubs directly from
-                    // each .scores.parquet listed via --input-scores. Skips
-                    // Stages 1-4. Side data (calibration, sidecars) is not
-                    // loaded on the C# side yet -- the simple Stage 5 path
-                    // doesn't need it. When reconciliation lands, this branch
-                    // will need to load the calibration JSON sibling files
-                    // (best-effort, like the Rust impl).
-                    // Guard: hash check against current --library and search params.
-                    // Aborts with a clear, file-named error if the operator points
-                    // the merge node at parquets from a different scoring run.
-                    string validationError = ParquetScoreCache.ValidateScoresParquetGroup(
-                        config.InputScores, config, Program.VERSION, LogWarning);
-                    if (validationError != null)
-                        throw new InvalidDataException(validationError);
-
-                    LogInfo(string.Format(
-                        "--join-only: loading {0} per-file score parquet(s)",
-                        config.InputScores.Count));
-                    for (int fileIdx = 0; fileIdx < config.InputScores.Count; fileIdx++)
-                    {
-                        string parquetPath = config.InputScores[fileIdx];
-                        string fileName = Path.GetFileNameWithoutExtension(parquetPath);
-                        if (fileName.EndsWith(".scores", StringComparison.Ordinal))
-                            fileName = fileName.Substring(0, fileName.Length - ".scores".Length);
-                        LogInfo(string.Format("===== Loading file {0}/{1}: {2} (from {3}) =====",
-                            fileIdx + 1, config.InputScores.Count, fileName, parquetPath));
-                        var stubs = ParquetScoreCache.LoadFdrStubsFromParquet(parquetPath);
-                        // Stage 5+ (Percolator SVM) requires the 21 PIN features
-                        // on each FdrEntry. Load them in lockstep with the stubs
-                        // and bind by row index (parquet rows are stable).
-                        var features = ParquetScoreCache.LoadPinFeaturesFromParquet(parquetPath);
-                        if (features.Count != stubs.Count)
-                        {
-                            throw new InvalidDataException(string.Format(
-                                "--join-only: parquet {0} has {1} stubs but {2} feature rows",
-                                parquetPath, stubs.Count, features.Count));
-                        }
-                        for (int j = 0; j < stubs.Count; j++)
-                            stubs[j].Features = features[j];
-                        LogInfo(string.Format("  Loaded {0} FDR stubs + features", stubs.Count));
-                        perFileEntries.Add(new KeyValuePair<string, List<FdrEntry>>(fileName, stubs));
-                        perFileParquetPaths[fileName] = parquetPath;
-
-                        // Best-effort calibration JSON load for Stage 6
-                        // reconciliation. Mirrors osprey/src/pipeline.rs:2573-2588.
-                        try
-                        {
-                            string parquetDir = Path.GetDirectoryName(Path.GetFullPath(parquetPath));
-                            if (parquetDir != null)
-                            {
-                                // fileName is the bare input stem (the trailing
-                                // ".scores" was stripped above), so combining it
-                                // with parquetDir yields the same input-stem path
-                                // ProcessFile passes to CalibrationPathForInput.
-                                string calStemPath = Path.Combine(parquetDir, fileName);
-                                string calPath = CalibrationIO.CalibrationPathForInput(calStemPath, parquetDir);
-                                if (File.Exists(calPath))
-                                {
-                                    var calParams = CalibrationIO.LoadCalibration(calPath);
-                                    if (calParams.RtCalibration != null && calParams.RtCalibration.ModelParams != null)
-                                    {
-                                        var mp = calParams.RtCalibration.ModelParams;
-                                        // Cross-impl JSON-decode parity check: append
-                                        // the raw arrays as loaded from the calibration
-                                        // JSON, gated by OSPREY_DUMP_CALIBRATION.
-                                        if (OspreyDiagnostics.DumpCalibration)
-                                        {
-                                            OspreyDiagnostics.WriteStage6CalibrationDump(
-                                                fileName, mp.LibraryRts, mp.FittedRts);
-                                        }
-                                        var rtCal = RTCalibration.FromModelParams(
-                                            mp.LibraryRts, mp.FittedRts, mp.AbsResiduals,
-                                            calParams.RtCalibration.ResidualSD);
-                                        perFileCalibrations[fileName] = rtCal;
-                                    }
-                                }
-                            }
-                        }
-                        catch (Exception ex)
-                        {
-                            LogWarning(string.Format("  Failed to load calibration for {0}: {1}", fileName, ex.Message));
-                        }
-                    }
-                    // Cross-impl JSON-decode parity short-circuit: after every
-                    // parquet's calibration JSON has been loaded and dumped, exit
-                    // if OSPREY_CALIBRATION_ONLY is set. Pairs with
-                    // OSPREY_DUMP_CALIBRATION.
-                    if (OspreyDiagnostics.CalibrationOnly)
-                        OspreyDiagnostics.ExitAfterDump(@"OSPREY_CALIBRATION_ONLY");
-                }
-                else if (config.InputFiles.Count == 1)
-                {
-                    // Single file: process directly (no parallel overhead)
-                    string inputFile = config.InputFiles[0];
-                    string fileName = Path.GetFileNameWithoutExtension(inputFile);
-                    LogInfo("");
-                    LogInfo(string.Format("===== Processing file 1/1: {0} =====", inputFile));
-                    var fileResult = ProcessFile(inputFile, fileName, fullLibrary, config, parquetFooterMetadata, perFileCalibrations);
-                    if (fileResult != null)
-                        perFileEntries.Add(new KeyValuePair<string, List<FdrEntry>>(fileName, fileResult));
-                }
-                else if (maxParallelFiles == 1)
-                {
-                    // Strictly sequential: one file at a time. Matches the
-                    // memory envelope of the single-file path while still
-                    // sharing the library load. Useful for 3-file Astral
-                    // runs that would OOM in parallel.
-                    LogInfo(string.Format(
-                        "[BENCH] OSPREY_MAX_PARALLEL_FILES=1 - processing {0} files sequentially",
-                        config.InputFiles.Count));
-                    for (int fileIdx = 0; fileIdx < config.InputFiles.Count; fileIdx++)
-                    {
-                        string inputFile = config.InputFiles[fileIdx];
-                        string fileName = Path.GetFileNameWithoutExtension(inputFile);
-                        LogInfo(string.Format("===== Processing file {0}/{1}: {2} =====",
-                            fileIdx + 1, config.InputFiles.Count, inputFile));
-                        var fileResult = ProcessFile(inputFile, fileName, fullLibrary, config, parquetFooterMetadata, perFileCalibrations);
-                        if (fileResult != null)
-                            perFileEntries.Add(new KeyValuePair<string, List<FdrEntry>>(fileName, fileResult));
-                    }
-                }
-                else
-                {
-                    // Multiple files: process in parallel, optionally
-                    // capped via OSPREY_MAX_PARALLEL_FILES.
-                    var parallelOpts = new ParallelOptions();
-                    if (maxParallelFiles > 1)
-                    {
-                        parallelOpts.MaxDegreeOfParallelism = maxParallelFiles;
-                        LogInfo(string.Format(
-                            "[BENCH] OSPREY_MAX_PARALLEL_FILES={0} - capping parallel file count",
-                            maxParallelFiles));
-                    }
-                    var fileResults = new ConcurrentDictionary<int, KeyValuePair<string, List<FdrEntry>>>();
-                    Parallel.For(0, config.InputFiles.Count, parallelOpts, fileIdx =>
-                    {
-                        string inputFile = config.InputFiles[fileIdx];
-                        string fileName = Path.GetFileNameWithoutExtension(inputFile);
-                        LogInfo(string.Format("===== Processing file {0}/{1}: {2} =====",
-                            fileIdx + 1, config.InputFiles.Count, inputFile));
-                        var fileResult = ProcessFile(inputFile, fileName, fullLibrary, config, parquetFooterMetadata, perFileCalibrations);
-                        if (fileResult != null)
-                            fileResults[fileIdx] = new KeyValuePair<string, List<FdrEntry>>(fileName, fileResult);
-                    });
-                    // Collect in original order
-                    for (int i = 0; i < config.InputFiles.Count; i++)
-                    {
-                        KeyValuePair<string, List<FdrEntry>> result;
-                        if (fileResults.TryGetValue(i, out result))
-                            perFileEntries.Add(result);
-                    }
-                }
-                swAllFiles.Stop();
-                LogInfo(string.Format("[TIMING] All files processed: {0:F1}s",
-                    swAllFiles.Elapsed.TotalSeconds));
-
-                // End-to-end (non-joinOnly) modes: populate perFileParquetPaths
-                // from config.InputFiles so Stage 6 reconciliation can locate
-                // each file's freshly-written .scores.parquet to lazy-load CWT
-                // candidates from. ProcessFile writes the parquet whenever
-                // parquetFooterMetadata != null (now always set in non-joinOnly mode).
-                if (!joinOnly)
-                {
-                    foreach (string inputFile in config.InputFiles)
-                    {
-                        string fileName = Path.GetFileNameWithoutExtension(inputFile);
-                        perFileParquetPaths[fileName] = ParquetScoreCache.GetScoresPath(inputFile);
-                    }
-                }
-
-                int totalScored = 0;
-                foreach (var kvp in perFileEntries)
-                    totalScored += kvp.Value.Count;
-
-                LogInfo("");
-                LogInfo(string.Format(
-                    "Coelution analysis complete. {0} total scored entries across {1} files",
-                    totalScored, nFiles));
-
-                // --join-at-pass=2: load the 1st-pass FDR scores sidecar for
-                // each file onto the freshly-loaded stubs. The sidecar
-                // carries the persisted SVM scores + q-values from the
-                // straight-through pipeline run that produced these
-                // reconciled parquets. Without this load, RunFirstPassProteinFdr
-                // and the compaction step (next) would see uninitialized
-                // entry.Score = 0 / q = 1 for every entry — every protein
-                // group would tie at score 0 and the picked-protein FDR
-                // would collapse. Mirrors Rust pipeline.rs:3823 sidecar
-                // load order (1st-pass first, then 2nd-pass after
-                // compaction).
-                if (config.ExpectReconciledInput)
-                {
-                    // Build a fileName -> synthetic input path map so we
-                    // can resolve sidecar paths via FdrScoresSidecar
-                    // (which expects an mzML-style stem, not a parquet
-                    // stem). config.InputFiles was synthesized earlier
-                    // by RescoreHydration.SyntheticInputFromParquet,
-                    // which strips the trailing ".scores" — Pass1Path
-                    // then builds <dir>/<file_name>.1st-pass.fdr_scores.bin.
-                    var inputByFileName = new Dictionary<string, string>();
-                    foreach (var inputFile in config.InputFiles)
-                        inputByFileName[Path.GetFileNameWithoutExtension(inputFile)] = inputFile;
-
-                    foreach (var kvp in perFileEntries)
-                    {
-                        string fileName = kvp.Key;
-                        var entries = kvp.Value;
-                        string inputFile;
-                        string sidecarPath = inputByFileName.TryGetValue(fileName, out inputFile)
-                            ? FdrScoresSidecar.Pass1Path(inputFile)
-                            : null;
-                        if (sidecarPath == null || !File.Exists(sidecarPath))
-                        {
-                            LogError(string.Format(
-                                "--join-at-pass=2: missing 1st-pass FDR sidecar for {0} " +
-                                "(expected at {1}). Re-run a straight-through pipeline to " +
-                                "produce the sidecar.",
-                                fileName, sidecarPath ?? "<unresolved>"));
-                            return 1;
-                        }
-                        if (!FdrScoresSidecar.TryRead(sidecarPath, entries,
-                                FdrScoresSidecar.Pass.FirstPass))
-                        {
-                            LogError(string.Format(
-                                "--join-at-pass=2: 1st-pass sidecar at {0} failed to load " +
-                                "(magic / version / pass-byte / count / size mismatch).",
-                                sidecarPath));
-                            return 1;
-                        }
-                    }
-                    LogInfo(string.Format(
-                        "--join-at-pass=2: loaded 1st-pass FDR sidecars for {0} file(s)",
-                        perFileEntries.Count));
-                }
-
-                if (perFileEntries.Count == 0 || totalScored == 0)
-                {
-                    LogWarning("No scored entries found. Cannot perform FDR control.");
-                    return 0;
-                }
-
-                // --no-join: stop here. Per-file `.scores.parquet` files are
-                // now on disk; a separate `--join-only` invocation (typically
-                // on a merge node) will pick them up and run Stage 5+.
-                if (config.NoJoin)
-                {
-                    LogInfo(string.Format(
-                        "--no-join: Stage 1-4 complete. {0} entries scored across {1} file(s). " +
-                        "Per-file `.scores.parquet` written next to each input mzML. " +
-                        "Skipping FDR and blib output.",
-                        totalScored, nFiles));
-                    return 0;
-                }
+                var fullLibrary = perFileScoringTask.FullLibrary;
+                var libraryById = perFileScoringTask.LibraryById;
+                var perFileEntries = perFileScoringTask.PerFileEntries;
+                var perFileCalibrations = perFileScoringTask.PerFileCalibrations;
+                var perFileParquetPaths = perFileScoringTask.PerFileParquetPaths;
 
                 // First-join phase: Stage 5 first-pass FDR + first-pass
                 // protein FDR + 1st-pass FDR sidecar persistence +
@@ -813,7 +422,7 @@ namespace pwiz.OspreySharp
         /// <summary>
         /// Process a single mzML file: load spectra, calibrate RT, score coelution.
         /// </summary>
-        private List<FdrEntry> ProcessFile(
+        internal List<FdrEntry> ProcessFile(
             string inputFile, string fileName,
             List<LibraryEntry> fullLibrary, OspreyConfig config,
             Dictionary<string, string> parquetFooterMetadata,
